Print real error reasons for proxy errors in debug mode
This commit was made following a suggestion in #29 due to the way
Mojo::UserAgent handles proxy errors. Long story short, when a
connection to the specified proxy server fails, Mojo::UserAgent
returns an opaque "Proxy connection failed" without returning
the real reason of failure. This can be problematic when debugging,
thus now this module extracts the real failure reason in debug mode.

From a technical standpoint, this attaches an on "start" handler on
the main `UserAgent` object which is invoked for each connection attempt
made by the user agent. This includes `CONNECT` requests, and when that
happens the module attaches an on "finish" handler to that connection
which allows to extract the real failure reason.